# packages/api-client/openapi.snapshot.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations": {
      "post": {
        "operationId": "OrganizationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Organization"
        ]
      },
      "get": {
        "operationId": "OrganizationController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Organization"
        ]
      }
    },
    "/organizations/{id}": {
      "get": {
        "operationId": "OrganizationController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Organization"
        ]
      },
      "patch": {
        "operationId": "OrganizationController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Organization"
        ]
      }
    },
...
    "/organizations/registration-requests": {
      "post": {
        "operationId": "RegistrationController_submitOrgRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitOrgRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationRequestEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Registration"
        ]
      },
      "get": {
        "operationId": "RegistrationController_listOrgRequests",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Registration"
        ]
      }
    },
...
    "/pos/sessions/open": {
      "post": {
        "operationId": "PosController_openSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosSessionEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/pos/sessions/{id}/start-sales": {
      "post": {
        "operationId": "PosController_startSales",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosSessionEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/pos/sessions/{id}": {
      "get": {
        "operationId": "PosController_getSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosSessionEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/pos/sessions/{id}/start-close": {
      "post": {
        "operationId": "PosController_startClose",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/pos/sessions/{id}/reconciliation": {
      "get": {
        "operationId": "PosController_getReconciliation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      },
      "post": {
        "operationId": "PosController_submitReconciliation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReconciliationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionReconciliationEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/pos/sessions/{id}/reconciliation/approve": {
      "post": {
        "operationId": "PosController_approveVariance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionReconciliationEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/pos/sessions/{id}/close": {
      "post": {
        "operationId": "PosController_finalizeClose",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PosSessionEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/pos/sales/checkout": {
      "post": {
        "operationId": "PosController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/pos/sales/{id}": {
      "get": {
        "operationId": "PosController_getSale",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/pos/sales/{id}/return": {
      "post": {
        "operationId": "PosController_processReturn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessReturnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
    "/pos/sales/{id}/exchange": {
      "post": {
        "operationId": "PosController_processExchange",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessExchangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Pos"
        ]
      }
    },
...
  },
  "info": {
    "title": "ERP API",
    "description": "OpenAPI 3 document for the ERP API. Regenerate the web SDK: `pnpm openapi:generate` (API must be reachable). Machine-readable spec: GET /docs-json",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "branch-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Branch-Id"
      },
      "request-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Request-Id"
      },
      "idempotency-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Idempotency-Key"
      }
    },
    "schemas": {
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "contactEmail": {
            "type": "string",
            "format": "email"
          },
          "contactPhone": {
            "type": "string",
            "maxLength": 30
          }
        },
        "required": [
          "name",
          "contactEmail"
        ]
      },
      "OrganizationEntity": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "mainBranchId": {
            "type": "string"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "SUSPENDED"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string",
            "description": "Tenant isolation key — every row belongs to exactly one organization."
          },
          "branchId": {
            "type": "string",
            "description": "Optional branch scope; null for org-wide records."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string",
            "description": "UUID of the user who created this record."
          }
        },
        "required": [
          "name",
          "contactEmail",
          "status",
          "id",
          "organizationId",
          "createdAt",
          "updatedAt",
          "createdBy"
        ]
      },
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "contactEmail": {
            "type": "string",
            "format": "email"
          },
          "contactPhone": {
            "type": "string",
            "maxLength": 30
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED"
            ]
          }
        }
      },
...
      "SubmitOrgRegistrationDto": {
        "type": "object",
        "properties": {
          "organizationName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "contactEmail": {
            "type": "string",
            "format": "email"
          },
          "contactPhone": {
            "type": "string",
            "maxLength": 30
          },
          "ownerName": {
            "type": "string",
            "minLength": 2
          },
          "ownerEmail": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "organizationName",
          "contactEmail",
          "ownerName",
          "ownerEmail"
        ]
      },
      "RegistrationRequestEntity": {
        "type": "object",
        "properties": {
          "type": {
            "enum": [
              "ORGANIZATION",
              "BRANCH"
            ],
            "type": "string"
          },
          "requestData": {
            "type": "object"
          },
          "status": {
            "enum": [
              "PENDING_APPROVAL",
              "APPROVED",
              "REJECTED",
              "RESUBMITTED"
            ],
            "type": "string"
          },
          "reviewedBy": {
            "type": "string"
          },
          "reviewedAt": {
            "format": "date-time",
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string",
            "description": "Tenant isolation key — every row belongs to exactly one organization."
          },
          "branchId": {
            "type": "string",
            "description": "Optional branch scope; null for org-wide records."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string",
            "description": "UUID of the user who created this record."
          }
        },
        "required": [
          "type",
          "requestData",
          "status",
          "id",
          "organizationId",
          "createdAt",
          "updatedAt",
          "createdBy"
        ]
      },
...
      "OpenSessionDto": {
        "type": "object",
        "properties": {
          "terminalId": {
            "type": "string",
            "format": "uuid"
          },
          "branchId": {
            "type": "string",
            "format": "uuid"
          },
          "openingCashAmount": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "branchId",
          "openingCashAmount"
        ]
      },
      "PosSessionEntity": {
        "type": "object",
        "properties": {
          "terminalId": {
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "ACTIVE_SALES",
              "CLOSING",
              "CLOSED"
            ],
            "type": "string"
          },
          "openedBy": {
            "type": "string"
          },
          "openedAt": {
            "format": "date-time",
            "type": "string"
          },
          "closedBy": {
            "type": "string"
          },
          "closedAt": {
            "format": "date-time",
            "type": "string"
          },
          "openingCashAmount": {
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string",
            "description": "Tenant isolation key — every row belongs to exactly one organization."
          },
          "branchId": {
            "type": "string",
            "description": "Optional branch scope; null for org-wide records."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string",
            "description": "UUID of the user who created this record."
          }
        },
        "required": [
          "status",
          "openedBy",
          "openedAt",
          "openingCashAmount",
          "id",
          "organizationId",
          "createdAt",
          "updatedAt",
          "createdBy"
        ]
      },
      "SubmitReconciliationDto": {
        "type": "object",
        "properties": {
          "actualCash": {
            "type": "number",
            "minimum": 0
          },
          "notes": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "actualCash"
        ]
      },
      "SessionReconciliationEntity": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "expectedCash": {
            "type": "number"
          },
          "actualCash": {
            "type": "number"
          },
          "variance": {
            "type": "number"
          },
          "varianceApproved": {
            "type": "boolean"
          },
          "approvedBy": {
            "type": "string"
          },
          "approvedAt": {
            "format": "date-time",
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string",
            "description": "Tenant isolation key — every row belongs to exactly one organization."
          },
          "branchId": {
            "type": "string",
            "description": "Optional branch scope; null for org-wide records."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string",
            "description": "UUID of the user who created this record."
          }
        },
        "required": [
          "sessionId",
          "expectedCash",
          "actualCash",
          "variance",
          "varianceApproved",
          "id",
          "organizationId",
          "createdAt",
          "updatedAt",
          "createdBy"
        ]
      },
      "CheckoutLineDto": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number",
            "minimum": 0.01
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0
          },
          "taxAmount": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "itemId",
          "locationId",
          "quantity",
          "unitPrice",
          "taxAmount"
        ]
      },
...
      "CheckoutDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "lines": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckoutLineDto"
            }
          },
          "payments": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckoutPaymentDto"
            }
          },
          "cashAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "revenueAccountId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "sessionId",
          "lines",
          "payments",
          "cashAccountId",
          "revenueAccountId"
        ]
      },
...
      "SaleEntity": {
        "type": "object",
        "properties": {
          "documentNumber": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "subtotal": {
            "type": "number"
          },
          "taxAmount": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "status": {
            "enum": [
              "COMPLETED",
              "RETURNED",
              "PARTIALLY_RETURNED"
            ],
            "type": "string"
          },
          "saleDate": {
            "format": "date-time",
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleLineEntity"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentEntity"
            }
          },
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string",
            "description": "Tenant isolation key — every row belongs to exactly one organization."
          },
          "branchId": {
            "type": "string",
            "description": "Optional branch scope; null for org-wide records."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string",
            "description": "UUID of the user who created this record."
          }
        },
        "required": [
          "documentNumber",
          "sessionId",
          "subtotal",
          "taxAmount",
          "totalAmount",
          "status",
          "saleDate",
          "lines",
          "payments",
          "id",
          "organizationId",
          "createdAt",
          "updatedAt",
          "createdBy"
        ]
      },
...
      "ProcessReturnDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000
          },
          "lines": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnLineDto"
            }
          },
          "cashAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "revenueAccountId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "sessionId",
          "reason",
          "lines",
          "cashAccountId",
          "revenueAccountId"
        ]
      },
      "ReturnEntity": {
        "type": "object",
        "properties": {
          "documentNumber": {
            "type": "string"
          },
          "originalSaleId": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "subtotal": {
            "type": "number"
          },
          "taxAmount": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "returnDate": {
            "format": "date-time",
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnLineEntity"
            }
          },
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string",
            "description": "Tenant isolation key — every row belongs to exactly one organization."
          },
          "branchId": {
            "type": "string",
            "description": "Optional branch scope; null for org-wide records."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string",
            "description": "UUID of the user who created this record."
          }
        },
        "required": [
          "documentNumber",
          "originalSaleId",
          "sessionId",
          "subtotal",
          "taxAmount",
          "totalAmount",
          "reason",
          "returnDate",
          "lines",
          "id",
          "organizationId",
          "createdAt",
          "updatedAt",
          "createdBy"
        ]
      },
...
    }
  }
}